% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigram_identify_function.R
\name{bigram_identify}
\alias{bigram_identify}
\title{bigram_identify}
\usage{
bigram_identify(text_data, freq_bigram)
}
\arguments{
\item{text_data}{The text data where you want to apply the bigram analysis.}

\item{freq_bigram}{The minimum frequency of bigrams included in the dataframe (0 will include all)}
}
\value{
Dataframe
}
\description{
This function identifies frequency of bigrams from text.
}
