% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clean_function.R
\name{pre_clean}
\alias{pre_clean}
\title{pre_clean}
\usage{
pre_clean(TEXT_DATA)
}
\arguments{
\item{TEXT_DATA}{some text data}
}
\value{
text data vector
}
\description{
Cleaning function for TEXT data - for both WOS & Scopus
}
