% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_replace_function.R
\name{text_replace}
\alias{text_replace}
\title{text_replace}
\usage{
text_replace(DATA)
}
\arguments{
\item{DATA}{some text data}
}
\value{
Text dataframe
}
\description{
This removes spaces between words that usually co-occur.
}
