% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigram_identify_function.R
\name{trigram_identify}
\alias{trigram_identify}
\title{trigram_identify}
\usage{
trigram_identify(text_data, freq_bigram)
}
\arguments{
\item{text_data}{The text data where you want to apply the trigram analysis.}

\item{freq_bigram}{The minimum frequency of trigrams included in the dataframe (0 will include all)}
}
\value{
Dataframe
}
\description{
This function identifies frequency of trigrams from text.
}
