% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOS_create_TD_function.R
\name{WOS_create_TD}
\alias{WOS_create_TD}
\title{WOS_create_TD}
\usage{
WOS_create_TD(DF, year)
}
\arguments{
\item{DF}{Web of Science dataframe}

\item{year}{year (the end year - all articles before the given year)}
}
\value{
Dataframe
}
\description{
This creates a new TEXT_DATA column for Web of Science DF for JOURNAL articles
}
