% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_read_function.R
\name{WOS_read}
\alias{WOS_read}
\title{WOS_read}
\usage{
WOS_read(pattern)
}
\arguments{
\item{pattern}{File pattern name (such as "savedrecs")}
}
\value{
Dataframe
}
\description{
This takes a list of excel files downloaded from Web of Science (WOS) and combines them into a single dataframe.
}
