% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_clean_function.R
\name{pre_clean}
\alias{pre_clean}
\title{pre_clean}
\usage{
pre_clean(DATA)
}
\arguments{
\item{DATA}{Dataframe with TEXT_DATA column}
}
\value{
text data vector
}
\description{
Cleaning function for TEXT data - for both WOS & Scopus
}
